refactor: avoid new Parser() every time
